- Moved industry to last step
